src/agents-api/pyproject.toml
[project]
name = "agents-api"
version = "1.0.0"
description = "Julep's backend API"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiobotocore>=2.21.0",
  "anyio>=4.4.0",
  "arrow>=1.3.0",
  "async-lru>=2.0.4",
  "beartype>=0.18.5",
  "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
  "environs>=10.3.0",
  "fastapi>=0.115.4",
  "fire>=0.5.0",
  "google-re2>=1.1.20240702",
  "gunicorn>=23.0.0",
  "httpx>=0.27.0",
  "jinja2schema>=0.1.4",
  "jinja2>=3.1.4",
  "jsonschema>=4.22.0",
  "langchain-core>=0.3.14",
  "larch-pickle~=1.4.3",
  "litellm>=1.74.9",
  "lz4>=4.3.3",
  "msgpack>=1.1.0",
  "numpy>=2.0.0,<2.1.0",
  "openfeature-sdk>=0.8.1",
  "pandas>=2.2.2",
  "prometheus-client>=0.21.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "pydantic-partial>=0.5.5",
  "pydantic[email]>=2.10.2",
  "python-box>=7.2.0",
  "scalar-fastapi>=1.0.3",
  "sentry-sdk[fastapi]>=2.13.0",
  "simpleeval>=0.9.13",
  "simsimd>=5.9.4",
  "spacy>=3.8.2",
  "sse-starlette>=2.1.3",
  "temporalio[opentelemetry]>=1.8",
  "tenacity>=9.0.0",
  "thefuzz>=0.22.1",
  "tiktoken>=0.7.0",
  "UnleashClient>=5.12.0",
  "uvicorn==0.29.0",
  "uvloop>=0.21.0",
  "xxhash>=3.5.0",
  "spacy-chunks>=0.0.2",
  "uuid7>=0.1.0",
  "asyncpg>=0.30.0",
  "unique-namer>=1.6.1",
  "langcodes>=3.5.0",
  "langchain-text-splitters>=0.3.2",
  "deep-translator>=1.11.4",
  "markdownify>=1.0.0",
  "markdown2>=2.5.3",
  "aiohttp>=3.11.13",
  "websockets>=15.0.1",
  "python-multipart>=0.0.20",
  "cryptography>=45.0.6",
  "backoff>=2.2.1",
  "pyjwt>=2.10.1",
  "sqlalchemy>=2.0.42",
  "redis>=6.4.0",
  "psycopg2-binary>=2.9.10",
  "fastapi-sso>=0.18.0",
]

[dependency-groups]
dev = [
    "datamodel-code-generator>=0.26.3",
    "ipython>=8.30.0",
    "ipywidgets>=8.1.5",
    "julep>=1.43.1",
    "jupyterlab>=4.3.1",
    "pip>=24.3.1",
    "poethepoet>=0.31.1",
    "pyjwt>=2.10.1",
    "pyright>=1.1.391",
    "pytype>=2024.10.11",
    "ruff>=0.9.0",
    "sqlvalidator>=0.0.20",
    "testcontainers[postgres,localstack]>=4.9.0",
    "ward>=0.68.0b0",
    "pyanalyze>=0.13.1",
    "autotyping>=24.9.0",
    "psycopg[binary]>=3.2.5", # only for use inside tests for now
    "ty>=0.0.0a8",
]

[tool.setuptools]
py-modules = [
    "agents_api",
]

[tool.ty.rules]
invalid-parameter-default = "warn"
invalid-argument-type = "warn"
unknown-argument = "warn"
invalid-type-form = "warn"
missing-argument = "warn"
unresolved-attribute = "warn"
invalid-assignment = "warn"
possibly-unbound-attribute = "warn"
no-matching-overload = "warn"
not-iterable = "warn"
unsupported-operator = "warn"

